+++ server/i18n/server/i18n.fr.js
@@ -1,0 +1,51 @@
+/** ****************************************************************************
+ *
+ * The French Translation.
+ * (contains the translations for the server messages)
+ *
+ * i18n.fr.js is just a literal object that contains a set of properties.
+ * It can't be instantiated.
+ *
+ * Private Functions:
+ *  . none,
+ *
+ *
+ * Public Static Methods:
+ *  . none,
+ *
+ *
+ *
+ * @namespace    -
+ * @dependencies none
+ * @exports      -
+ * @author       -
+ * @since        0.0.0
+ * @version      -
+ * ************************************************************************** */
+/* eslint-disable no-underscore-dangle */
+
+
+// -- Vendor Modules
+
+
+// -- Local Modules
+
+
+// -- Local Constants
+
+
+// -- Local Variables
+
+
+// -- Public -------------------------------------------------------------------
+
+const French = {
+  Hello: 'Bonjour',
+  'Hello %s and %s': 'Bonjour %s et %s',
+};
+
+
+// -- Export
+module.exports = French;
+
+/* - */