+++ server/start.js
@@ -17,13 +17,16 @@
  * @since     0.0.0
  * @version   -
  * ************************************************************************** */
+/* eslint one-var: 0, semi-style: 0, no-underscore-dangle: 0 */
 
 
 // -- Vendor Modules
 
 
 // -- Local Modules
-const App = require('./app');
+const App = require('./app')
+    , DB  = require('./db/build')
+    ;
 
 
 // -- Local Constants
@@ -34,4 +37,11 @@ const App = require('./app');
 
 // -- Main section -
 
+// This is for testing purpose. It initializes the SQlite database
+// './db/db.sqlite' with credential for the user authentication.
+// DB.init();
+
+// Starts the server:
 App();
+
+/* - */