+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ### HEAD
 
+### 1.0.3 (December 22, 2020)
+
+  * Added internationalization capability through a new i18n local module,
+  * Added an api to access to an sqlite3 database,
+  * Stored the credentials into the sqlite db with hashed passwords,
+  * ...,
+
+
 ### 1.0.2 (October 16, 2020)
 
   * Added the script tasks/prod.sh to build a production version,